Merged css->html (avoid link dep.)
Moved the css within the head of the html to avoid issues
with linking dependencies (i.e. that outfile, depending on where
it is output to, will not be able to find the css unless this
is swithed to abs path)

Also, added some default values to invoice.html to clean up
template.